change float to double

+++ INB372-Solar/src/com/inb372/solar/trariffCalculation.java
@@ -5,8 +5,8 @@ package com.inb372.solar;
 public class trariffCalculation {
 
 	private Calculator solar;
-	final double cost=0.44;
-	double electric;
+	final float cost=0.44;
+	float electric;
 	//
 	public trariffCalculation(Calculator solar) throws trariffException {
 		if (solar.calcDailyExcess() < 0)
@@ -15,12 +15,12 @@ public class trariffCalculation {
 	}
 	
 	//Calculate the Annual Electric
-	public double calAnnualElectric() {	
+	public float calAnnualElectric() {	
 		return electric * 365;
 	}
 	
 	//Calculate the Annual Saving
-	public double calAnnualSaving() {
+	public float calAnnualSaving() {
 		return calAnnualElectric() * cost;
 	}
 }
